pyproject.toml
[project]
name = "tg-finance"
version = "0.1.0"
description = "Telegram bot for financial accounting."
authors = [{ name = "Dmitry Pushkin" }]
requires-python = "~=3.12"
readme = "README.md"
dependencies = [
    "pydantic>=2.10.6,<3",
    "pydantic-settings>=2.7.1,<3",
    "sqlalchemy[asyncio]>=2.0.37,<3",
    "asyncpg>=0.30.0,<0.31",
    "alembic>=1.14.1,<2",
    "orjson>=3.10.15,<4",
    "alembic-postgresql-enum>=1.6.0,<2",
    "python-dateutil>=2.9.0.post0,<3",
    "pytz>=2025.1",
    "python-telegram-bot[job-queue]>=22.0",
    "redis>=5.2.1",
]

[tool.uv]
package = false
